+++ generators/typescript/sdk/client-class-generator/src/SdkClientClassGenerator.ts
@@ -20,7 +20,7 @@ export declare namespace SdkClientClassGenerator {
         includeContentHeadersOnFileDownloadResponse: boolean;
         includeSerdeLayer: boolean;
         retainOriginalCasing: boolean;
-        wrapFileProperties: boolean;
+        inlineFileProperties: boolean;
         oauthTokenProviderGenerator: OAuthTokenProviderGenerator;
     }
 
@@ -48,7 +48,7 @@ export class SdkClientClassGenerator {
     private includeContentHeadersOnFileDownloadResponse: boolean;
     private includeSerdeLayer: boolean;
     private retainOriginalCasing: boolean;
-    private wrapFileProperties: boolean;
+    private inlineFileProperties: boolean;
     private oauthTokenProviderGenerator: OAuthTokenProviderGenerator;
 
     constructor({
@@ -65,7 +65,7 @@ export class SdkClientClassGenerator {
         includeContentHeadersOnFileDownloadResponse,
         includeSerdeLayer,
         retainOriginalCasing,
-        wrapFileProperties,
+        inlineFileProperties,
         oauthTokenProviderGenerator
     }: SdkClientClassGenerator.Init) {
         this.intermediateRepresentation = intermediateRepresentation;
@@ -81,7 +81,7 @@ export class SdkClientClassGenerator {
         this.includeContentHeadersOnFileDownloadResponse = includeContentHeadersOnFileDownloadResponse;
         this.includeSerdeLayer = includeSerdeLayer;
         this.retainOriginalCasing = retainOriginalCasing;
-        this.wrapFileProperties = wrapFileProperties;
+        this.inlineFileProperties = inlineFileProperties;
         this.oauthTokenProviderGenerator = oauthTokenProviderGenerator;
     }
 
@@ -109,7 +109,7 @@ export class SdkClientClassGenerator {
             includeContentHeadersOnFileDownloadResponse: this.includeContentHeadersOnFileDownloadResponse,
             includeSerdeLayer: this.includeSerdeLayer,
             retainOriginalCasing: this.retainOriginalCasing,
-            wrapFileProperties: this.wrapFileProperties,
+            inlineFileProperties: this.inlineFileProperties,
             oauthTokenProviderGenerator: this.oauthTokenProviderGenerator
         });
     }