+++ generators/typescript/utils/contexts/src/sdk-context/SdkContext.ts
@@ -29,5 +29,5 @@ export interface SdkContext extends ModelContext {
     includeSerdeLayer: boolean;
     retainOriginalCasing: boolean;
     generateOAuthClients: boolean;
-    wrapFileProperties: boolean;
+    inlineFileProperties: boolean;
 }